Update from upstream repo (@ilkon) by @enchance
Integrated updates from the upstream repo of which @enchance was made a
contributor of. These updates include the ff in order of `git log`:

- Switch to InnoDB for better data integrity, transactions, and FK (6952d9)
- Remove unused methods (#50) (84d869)
- Create generate_random_key() c/o @inda5th (deb24a)
- Remove notice controller (fixes #46) (05b016)
- Replace rand() with mt_rand() (d25e4e)
- Add country schema (11d320)
- Update schema (f48f56)
- Update library methods and added comments (cb29b1)

I added these updates since work on upstream might stay in the devel
branch for some time prompting me to use this fork for an extended period.